backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database (PostgreSQL)
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Configuration
pydantic-settings==2.5.2
python-dotenv==1.0.1

# NLP & ML
spacy==3.7.6
sentence-transformers==3.1.0
faiss-cpu==1.8.0
pgvector==0.3.2
scikit-learn==1.5.2
numpy==1.26.4
langdetect==1.0.9
Unidecode==1.4.0

# RSS & HTTP
feedparser==6.0.11
httpx==0.27.2

# AWS
boto3==1.35.0

# Utilities
python-dateutil==2.9.0
apscheduler==3.11.2

# Caching
redis>=5.0.0

# Full content scraping
trafilatura>=1.6.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
